extract utils into extensions
