+++ server/extensions.js
@@ -1,0 +1,6 @@
+'use strict';
+
+Array.prototype.concatMap = function (f) {
+  return Array.prototype.concat.apply([], this.map(f));
+};
+
