+++ test/expectations.test.js
@@ -131,12 +131,3 @@ describe('hasUsage', function () {
   });
 
 });
-
-describe('utils', function () {
-
-  it('concatMap', function () {
-    var concatMap = [{foo: [1,2]}, {foo: [3,4]}].concatMap(function (obj) { return obj.foo; });
-    concatMap.should.be.eql([1,2,3,4]);
-  });
-
-});
